+++ R/tableConceptIdCounts.R
@@ -11,11 +11,14 @@ tableConceptIdCounts <- function(result,
   rlang::check_installed("visOmopResults")
   omopgenerics::validateResultArgument(result)
   omopgenerics::assertChoice(type, visOmopResults::tableType())
-
+  strata_cols <- omopgenerics::strataColumns(result)
   # subset to result_type of interest
   result <- result |>
     omopgenerics::filterSettings(
-      .data$result_type == "summarise_concept_id_counts")
+      .data$result_type == "summarise_concept_id_counts")|>
+    omopgenerics::splitStrata() |>
+    dplyr::arrange(.data$variable_name, dplyr::across(dplyr::all_of(strata_cols)), .data$additional_level)|>
+    omopgenerics::uniteStrata(strata_cols)
 
   # check if it is empty
   if (nrow(result) == 0) {
@@ -27,24 +30,26 @@ tableConceptIdCounts <- function(result,
     dplyr::distinct(.data$estimate_name) |>
     dplyr::pull()
   estimateName <- c()
-  if ("record_count" %in% estimate_names) {
-    estimateName <- c(estimateName, "N records" = "<record_count>")
+  if ("count_records" %in% estimate_names) {
+    estimateName <- c(estimateName, "N records" = "<count_records>")
   }
-  if ("person_count" %in% estimate_names) {
-    estimateName <- c(estimateName, "N persons" = "<person_count>")
+  if ("count_subjects" %in% estimate_names) {
+    estimateName <- c(estimateName, "N persons" = "<count_subjects>")
   }
-  if (type == "datatable" && dplyr::n_distinct(result$cdm_name) == 1) {
-    header <- NULL
-  } else {
-    header <- c("cdm_name")
-  }
+ header <- c("cdm_name", "estimate_name")
+
   result |>
     formatColumn(c("variable_name", "variable_level")) |>
     visOmopResults::visOmopTable(
       type = type,
       estimateName = estimateName,
         header = header,
-        rename = c("Database name" = "cdm_name"),
-        groupColumn = c("omop_table", omopgenerics::strataColumns(result))
+        rename = c("Database name" = "cdm_name", "Concept name" = "variable_name", "Concept id" = "variable_level" ),
+        groupColumn = c(omopgenerics::additionalColumns(result)),
+      columnOrder = c("omop_table","variable_name", "variable_level", strata_cols),
+
+
+      .options = list(groupAsColumn = TRUE, merge = "all_columns"
+      )
       )
 }